+++ packages/mint-components/CHANGELOG.md
@@ -24,7 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   /(rewardsCountFiltered)/:statType?/:unit?/:status([PENDING|CANCELLED|EXPIRED|REDEEMED|AVAILABLE]*)?/:global?
   /(integrationRewardsCountFiltered)/:status([PENDING|CANCELLED|EXPIRED|REDEEMED|AVAILABLE]*)?/:global?
   ```
-- <sqm-share-button> loading states
+- <sqm-share-button> loading states and hide text prop
 - <sqm-divided-layout> border customizable
 
 ## [1.2.0] - 2021-08-31